test(subscriptions): extract inline tests to sibling files
Per project rule #13 (separate test files) and rule #14 (modernize touched
modules. Each implementation file now wires its sibling via:

  #[cfg(...)] #[path = <name>_tests.rs] mod tests;

- subscriptions/sync.rs       -> sync_tests.rs
- subscriptions/async.rs      -> async_tests.rs
- subscriptions/common.rs     -> common_tests.rs

No test changes; pure file relocation.
EOF
)
